+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## TO BE RELEASED
 
+* enable RDS performance insights
 * MI-125: rename maven cache file
 * MI-127: remove crowdstrike install from runlists
 * MI-130: `PermissionsSyncer` now compares configured user `ssh_key` value with existing and updates if they differ